pyproject.toml
[tool.poetry]
name = "image-converter"
version = "0.1.0"
description = ""
authors = ["Phillip Strefling <phillip@mindblox.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
heic-to-jpg = "^0.1.9"
google-api-python-client = "^2.111.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
pillow = "^10.2.0"
pyheif = "^0.7.1"
fire = "^0.5.0"
tqdm = "^4.66.1"
pytest-mock = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
convert = "image_converter.app:run"

